@@ -17,7 +17,7 @@ use RestOauth\Test\Fixture\OauthClientsFixture;
 
 class OauthTokenControllerTest extends ApiCommonErrorsTest
 {
-    protected $fixtures = [
+    protected array $fixtures = [
         UsersFixture::LOAD,
         OauthClientsFixture::LOAD,
     ];
